[frontend] fix AI Insights display in entity overviews (#16116)
