+++ opencti-platform/opencti-front/src/private/components/arsenal/malwares/Root.tsx
@@ -21,6 +21,7 @@ import ErrorNotFound from '../../../../components/ErrorNotFound';
 import { useFormatter } from '../../../../components/i18n';
 import Breadcrumbs from '../../../../components/Breadcrumbs';
 import { getPaddingRight } from '../../../../utils/utils';
+import { isPathOverview } from '../../../../utils/tabUtils';
 import { RootMalwareQuery } from './__generated__/RootMalwareQuery.graphql';
 import { RootMalwareSubscription } from './__generated__/RootMalwareSubscription.graphql';
 import Security from '../../../../utils/Security';
@@ -101,7 +102,7 @@ const RootMalware = ({ queryRef, malwareId }: RootMalwareProps) => {
   } = usePreloadedQuery<RootMalwareQuery>(malwareQuery, queryRef);
   const { forceUpdate } = useForceUpdate();
   const basePath = PATH_MALWARE(malwareId);
-  const isOverview = location.pathname === basePath;
+  const isOverview = isPathOverview(location.pathname, basePath);
   const paddingRight = getPaddingRight(location.pathname, basePath);
   const link = `${basePath}/knowledge`;
   return (
