+++ opencti-platform/opencti-front/src/private/components/entities/sectors/Root.tsx
@@ -24,6 +24,7 @@ import EntityStixSightingRelationships from '../../events/stix_sighting_relation
 import { useFormatter } from '../../../../components/i18n';
 import Breadcrumbs from '../../../../components/Breadcrumbs';
 import { getPaddingRight } from '../../../../utils/utils';
+import { isPathOverview } from '../../../../utils/tabUtils';
 import Security from '../../../../utils/Security';
 import { KNOWLEDGE_KNUPDATE, KNOWLEDGE_KNUPDATE_KNDELETE } from '../../../../utils/hooks/useGranted';
 import SectorEdition from './SectorEdition';
@@ -102,7 +103,7 @@ const RootSector = ({ sectorId, queryRef }: RootSectorProps) => {
   const { forceUpdate } = useForceUpdate();
 
   const basePath = PATH_SECTOR(sectorId);
-  const isOverview = location.pathname === basePath;
+  const isOverview = isPathOverview(location.pathname, basePath);
   const paddingRight = getPaddingRight(location.pathname, basePath);
   const link = `${basePath}/knowledge`;
   return (
