+++ opencti-platform/opencti-front/src/private/components/threats/threat_actors_individual/Root.tsx
@@ -21,6 +21,7 @@ import StixCoreObjectKnowledgeBar from '../../common/stix_core_objects/StixCoreO
 import { useFormatter } from '../../../../components/i18n';
 import Breadcrumbs from '../../../../components/Breadcrumbs';
 import { getPaddingRight } from '../../../../utils/utils';
+import { isPathOverview } from '../../../../utils/tabUtils';
 import Security from '../../../../utils/Security';
 import { KNOWLEDGE_KNUPDATE, KNOWLEDGE_KNUPDATE_KNDELETE } from '../../../../utils/hooks/useGranted';
 import ThreatActorIndividualEdition from './ThreatActorIndividualEdition';
@@ -113,7 +114,7 @@ const RootThreatActorIndividualComponent = ({
   );
   const { forceUpdate } = useForceUpdate();
   const basePath = PATH_THREAT_ACTORS_INDIVIDUAL(threatActorIndividualId);
-  const isOverview = location.pathname === basePath;
+  const isOverview = isPathOverview(location.pathname, basePath);
   const paddingRight = getPaddingRight(location.pathname, basePath);
   const link = `${basePath}/knowledge`;
   return (
