+++ opencti-platform/opencti-front/src/utils/utils.test.ts
@@ -1,31 +1,6 @@
 import { describe, expect, it } from 'vitest';
-import { getCurrentTab, getPaddingRight } from './utils';
+import { getPaddingRight } from './utils';
 
-describe('getCurrentTab', () => {
-  it('returns next segment after the basepath', () => {
-    const value = getCurrentTab(
-      '/dashboard/techniques/attack_patterns/965f2989-6acb-4223-a163-89a99411c15c/knowledge',
-      '/dashboard/techniques/attack_patterns/965f2989-6acb-4223-a163-89a99411c15c',
-    );
-    expect(value).toBe('knowledge');
-  });
-
-  it('returns an empty string when fullpath equals basepath', () => {
-    const value = getCurrentTab(
-      '/dashboard/techniques/attack_patterns/965f2989-6acb-4223-a163-89a99411c15c',
-      '/dashboard/techniques/attack_patterns/965f2989-6acb-4223-a163-89a99411c15c',
-    );
-    expect(value).toBe('');
-  });
-
-  it('returns next segment after the basepath even when there are more', () => {
-    const value = getCurrentTab(
-      '/dashboard/techniques/attack_patterns/965f2989-6acb-4223-a163-89a99411c15c/knowledge/or/something/else',
-      '/dashboard/techniques/attack_patterns/965f2989-6acb-4223-a163-89a99411c15c',
-    );
-    expect(value).toBe('knowledge');
-  });
-});
 
 describe('getPaddingRight', () => {
   const basePath = '/dashboard/analyses/reports/a079e1e7-ce97-4528-92f1-64df365d540b';
