+++ opencti-platform/opencti-front/src/utils/utils.ts
@@ -46,33 +46,6 @@ export const getFileUri = (id: string) => {
 export const uniqueArray = <T>(items: IterableIterator<T> | Array<T>) => Array.from(new Set(items));
 
 /**
- * Compute the current state value of a <Tabs> component by comparing the
- * location.pathname (`fullpath`) with the `basePath` of the component
- * and extracting the next URL segment in the sequence.
- *
- * @param fullpath - The current full pathname as returned by useLocation().pathname.
- * Should contain no query params nor hash param.
- * @param basePath - The closest path to root where the <Tabs> component is rendered.
- * @returns The current tab value.
- *
- * @example
- * ```
- * getCurrentTab(
- *   '/dashboard/techniques/attack_patterns/965f2989-6acb-4223-a163-89a99411c15c/knowledge/or/something/else',
- *   '/dashboard/techniques/attack_patterns/965f2989-6acb-4223-a163-89a99411c15c',
- * ); // returns 'knowledge'
- * ```
- */
-export const getCurrentTab = (fullpath: string, basePath: string) => {
-  let subpath = fullpath.substring(basePath.length);
-  if (subpath.startsWith('/')) {
-    subpath = subpath.substring(1);
-  }
-  const nextSlashPos = subpath.indexOf('/');
-  return nextSlashPos >= 0 ? subpath.substring(0, nextSlashPos) : subpath;
-};
-
-/**
  * Compute the right padding to apply on an entity page based on the current route.
  * - Knowledge pages get extra padding (200) for the timeline/graph side panel.
  * - Content pages get padding (350) for the files side panel, except the mapping view which needs full width.
